+++ src/main/scala/s3/website/CloudFront.scala
@@ -121,7 +121,7 @@ object CloudFront {
 
 
   def needsInvalidation: PartialFunction[PushSuccessReport, Boolean] = {
-    case SuccessfulUpload(upload, _) => upload.uploadType match {
+    case SuccessfulUpload(upload, _, _) => upload.uploadType match {
       case Update => true
       case _ => false
     }